Quick Fix for cannot find host node
In the actual react-router version with the Link Module this is not working anymore. Console is giving a:
http://stackoverflow.com/questions/38796376/cannot-read-property-gethostnode-of-null

Actually you removing some features, since a simple a tag does always render the whole page new and doesnt support a transition. So pls just take this as a hint about what to do. Otherwise the react-router part had to be fixed. The comments are the problem.